Added a GameContainer component to hold the rooms API call. Gamescreen now renders rooms after receiving a prop of the API response
