+++ adventure/src/App.js
@@ -4,7 +4,7 @@ import Login from './components/Login';
 import Register from './components/Register';
 import PrivateRoute from './components/PrivateRoute';
 import {BrowserRouter as Router, Route, Link} from 'react-router-dom';
-import GameScreen from './components/GameScreen';
+import GameContainer from './components/GameContainer';
 
 function App() {
   return (
@@ -18,7 +18,7 @@ function App() {
 
       <Route path="/login" component={Login} />
       <Route path="/register" component={Register} />
-      <PrivateRoute path="/dashboard" component={GameScreen} />
+      <PrivateRoute path="/dashboard" component={GameContainer} />
     </div>
     </Router>
   );
